I believe that I may have finished the entire screen but I haven't ran it yet, but I'll do that later
